Fix: 해쉬태그 오류 수정, 컴포넌트 분리
